final commit added comments and tests

+++ src/main/java/com/Mustafa/OtherOptions.java
@@ -1,7 +1,7 @@
 package com.Mustafa;
 
 import java.io.*;
-import java.lang.reflect.Array;
+import java.util.TreeSet;
 
 public class OtherOptions extends MainMenu {
 
@@ -10,7 +10,7 @@ public class OtherOptions extends MainMenu {
         try {
             // Add file not a case sensitive
             scanner.nextLine();// Clean scanner cash .
-            System.out.println("Please add the source file's path to add (Example:'C:/hcl/learning/test.txt'):");
+            System.out.println("Please enter the source file's path to add (Example:'C:/hcl/learning/test.txt'):");
             String resourcePath = scanner.nextLine();// Get the path to copy from .
             File f = new File(resourcePath);
             String filename = f.getName(); // Extract the file name.
@@ -37,32 +37,66 @@ public class OtherOptions extends MainMenu {
         mm.mainMenuOptions();
 
     }
-
+    // Delete a file case sensitive.
     static void deleteFile() {
+        boolean flag = false;
 
         scanner.nextLine();// Clean scanner cash .
         System.out.println("Please Enter a file name to delete Example (Test.txt):");
-        String fileToDelete = scanner.nextLine();// Get the path to copy from .
-        File f = new File(path + "/" + fileToDelete);
-        boolean result = f.delete();
-        if (result) {
-            System.out.println("File name ' " + fileToDelete + " ' has been deleted successfully");
-        } else {
-            System.out.println("File not found.");
-        }
-    }
-
-    static void searchFile() {
-
-        scanner.nextLine();// Clean scanner cash .
-        System.out.println("Please Enter a file name to search for Example (Test.txt):");
-        String fileToDelete = scanner.nextLine();// Get the path to copy from .
+        String fileToDelete = scanner.nextLine();// Get the file name form the user to delete  .
         File f = new File(path + "/" + fileToDelete);
         boolean result = f.exists();// Check if the file exist.
         if (result) {
-            System.out.println("File name ' " + fileToDelete + " ' has been found successfully");
+            TreeSet<String> setOfFiles = new TreeSet();
+            File[] files = new File(path).listFiles();
+            for (File file : files) {
+                if (file.isFile()) {
+                    if (file.getName().equals(fileToDelete)) { //Case sensitivity file name it should be exactly same to delete.
+                        boolean delResult = f.delete();
+                        if (delResult) {
+                            flag = true;
+                            System.out.println("File name ' " + fileToDelete + " ' has been deleted successfully");
+                            break;
+                        }
+
+                    }
+                }
+            }
+            if (!flag) {
+                System.out.println("File not found!.");
+            }
+
         } else {
-            System.out.println("File not found .");
+            System.out.println("File not found!.");
+        }
+    }
+    // Search for a file case sensitive :
+    static void searchFile() {
+        boolean flag = false;
+        scanner.nextLine();// Clean scanner cash .
+        System.out.println("Please Enter a file name to search for Example (Test.txt):");
+        String fileToSearch = scanner.nextLine();// Get file name to search for .
+        File f = new File(path + "/" + fileToSearch);
+        boolean result = f.exists();// Check if the file exist.
+
+        if (result) { // If the file exists then check the case sensitivity
+            TreeSet<String> setOfFiles = new TreeSet();
+            File[] files = new File(path).listFiles();
+            for (File file : files) {
+                if (file.isFile()) {
+                    if (file.getName().equals(fileToSearch)) { //Case sensitivity it should be exactly same to print the founded file.
+                        flag = true;
+                        System.out.println("File name ' " + fileToSearch + " ' has been found successfully");
+                        break;
+                    }
+                }
+            }
+            if (!flag) {
+                System.out.println("File not found! .");
+            }
+
+        } else {
+            System.out.println("File not found! .");
         }
 
     }
